+++ src/supabaseClient.js
@@ -1,0 +1,8 @@
+const { createClient } = require("@supabase/supabase-js");
+require("dotenv").config();
+
+const supabaseUrl = process.env.SUPABASE_URL;
+const supabaseKey = process.env.SUPABASE_KEY;
+const supabase = createClient(supabaseUrl, supabaseKey);
+
+module.exports = { supabase };
