+++ vercel.json
@@ -2,7 +2,7 @@
   "version": 2,
   "builds": [
     {
-      "src": "./index.js",
+      "src": ".src/server.js",
       "use": "@vercel/node"
     }
   ],
